# packages/contracts/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Aura local foundation",
    "version": "0.0.1"
  },
  "paths": {
...
    "/api/jobs": {
      "get": {
        "summary": "List Jobs",
        "operationId": "list_jobs_api_jobs_get",
        "parameters": [
          {
            "name": "state",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "$ref": "#/components/schemas/JobState"
                },
                {
                  "type": "null"
                }
              ],
              "title": "State"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/JobRecord"
                  },
                  "title": "Response List Jobs Api Jobs Get"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/jobs/{job_id}/redrive": {
      "post": {
        "summary": "Redrive",
        "description": "Requeue a dead letter with its identity; audited as JOB_REDRIVEN in the outbox.",
        "operationId": "redrive_api_jobs__job_id__redrive_post",
        "parameters": [
          {
            "name": "job_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Job Id"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/RedriveCommand"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/JobCommandResult"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "JobCommandResult": {
        "properties": {
          "job_id": {
            "type": "string",
            "title": "Job Id"
          },
          "version": {
            "type": "integer",
            "title": "Version"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "job_id",
          "version"
        ],
        "title": "JobCommandResult",
        "description": "The job's version after an audited command, or the recorded one on replay."
      },
      "JobRecord": {
        "properties": {
          "id": {
            "type": "string",
            "title": "Id"
          },
          "kind": {
            "type": "string",
            "title": "Kind"
          },
          "job_class": {
            "$ref": "#/components/schemas/WorkClass"
          },
          "scope": {
            "type": "string",
            "title": "Scope"
          },
          "occurrence_key": {
            "type": "string",
            "title": "Occurrence Key"
          },
          "correlation_id": {
            "type": "string",
            "title": "Correlation Id"
          },
          "payload": {
            "additionalProperties": true,
            "type": "object",
            "title": "Payload"
          },
          "input_version": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Input Version"
          },
          "priority": {
            "type": "integer",
            "title": "Priority"
          },
          "state": {
            "$ref": "#/components/schemas/JobState"
          },
          "version": {
            "type": "integer",
            "title": "Version"
          },
          "attempts": {
            "type": "integer",
            "title": "Attempts"
          },
          "max_attempts": {
            "type": "integer",
            "title": "Max Attempts"
          },
          "redrives": {
            "type": "integer",
            "title": "Redrives"
          },
          "fencing_token": {
            "type": "integer",
            "title": "Fencing Token"
          },
          "lease_owner": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Lease Owner"
          },
          "lease_expires_at": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Lease Expires At"
          },
          "scheduled_at": {
            "type": "string",
            "title": "Scheduled At"
          },
          "available_at": {
            "type": "string",
            "title": "Available At"
          },
          "deadline": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Deadline"
          },
          "terminal_reason": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Terminal Reason"
          },
          "last_error": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Last Error"
          },
          "result": {
            "anyOf": [
              {
                "additionalProperties": true,
                "type": "object"
              },
              {
                "type": "null"
              }
            ],
            "title": "Result"
          },
          "created_at": {
            "type": "string",
            "title": "Created At"
          },
          "updated_at": {
            "type": "string",
            "title": "Updated At"
          },
          "finished_at": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Finished At"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "id",
          "kind",
          "job_class",
          "scope",
          "occurrence_key",
          "correlation_id",
          "payload",
          "input_version",
          "priority",
          "state",
          "version",
          "attempts",
          "max_attempts",
          "redrives",
          "fencing_token",
          "lease_owner",
          "lease_expires_at",
          "scheduled_at",
          "available_at",
          "deadline",
          "terminal_reason",
          "last_error",
          "result",
          "created_at",
          "updated_at",
          "finished_at"
        ],
        "title": "JobRecord",
        "description": "A durable job. Dead letters are jobs in state DEAD; a redrive keeps id and keys."
      },
      "JobState": {
        "type": "string",
        "enum": [
          "QUEUED",
          "RUNNING",
          "SUCCEEDED",
          "DEAD",
          "EXPIRED"
        ],
        "title": "JobState"
      },
...
      "RedriveCommand": {
        "properties": {
          "expected_version": {
            "type": "integer",
            "minimum": 1.0,
            "title": "Expected Version"
          },
          "reason": {
            "type": "string",
            "maxLength": 500,
            "minLength": 4,
            "title": "Reason"
          },
          "command_id": {
            "type": "string",
            "format": "uuid",
            "title": "Command Id"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "expected_version",
          "reason",
          "command_id"
        ],
        "title": "RedriveCommand"
      },
...
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          },
          "input": {
            "title": "Input"
          },
          "ctx": {
            "type": "object",
            "title": "Context"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      },
      "WorkClass": {
        "type": "string",
        "enum": [
          "RECONCILIATION",
          "MONITORING",
          "INGEST",
          "ANALYSIS",
          "RESEARCH"
        ],
        "title": "WorkClass",
        "description": "Chapter 08 work classes. Reconciliation and monitoring are safety work."
      }
    }
  }
}